requirements.txt
Flask==2.3.3
python-dotenv==1.0.0
gspread==5.12.0
google-auth==2.23.0
Flask-Session==0.5.0
gunicorn==21.2.0
requests==2.31.0
gevent==23.9.1
Pillow==10.0.0